Refactor. Added pathfinder base class. moved current simple pathfiding logic to a subclass of pathfinder.
